requirements.txt
streamlit==1.28.1
llama-cpp-python==0.2.11
huggingface-hub==0.17.3
requests==2.31.0
pydantic==2.4.2